Separate client to module
